Separate BaseOperationModel into its own module'
