+++ floris/core/turbine/__init__.py
@@ -1,12 +1,11 @@
-
-from floris.core.turbine.controller_dependent_operation_model import ControllerDependentTurbine
+from floris.core.turbine.operation_model_base import BaseOperationModel
 from floris.core.turbine.operation_models import (
     AWCTurbine,
-    BaseOperationModel,
     CosineLossTurbine,
     MixedOperationTurbine,
     PeakShavingTurbine,
     SimpleDeratingTurbine,
     SimpleTurbine,
 )
+from floris.core.turbine.controller_dependent_operation_model import ControllerDependentTurbine
 from floris.core.turbine.unified_momentum_model import UnifiedMomentumModelTurbine
