+++ floris/core/turbine/controller_dependent_operation_model.py
@@ -10,7 +10,7 @@ from floris.core.rotor_velocity import (
     compute_tilt_angles_for_floating_turbines,
     rotor_velocity_air_density_correction,
 )
-from floris.core.turbine.operation_models import BaseOperationModel
+from floris.core.turbine import BaseOperationModel
 from floris.type_dec import (
     NDArrayFloat,
     NDArrayObject,
