+++ floris/core/turbine/operation_models.py
@@ -18,6 +18,7 @@ from floris.core.rotor_velocity import (
     rotor_velocity_tilt_cosine_correction,
     rotor_velocity_yaw_cosine_correction,
 )
+from floris.core.turbine import BaseOperationModel
 from floris.type_dec import (
     NDArrayFloat,
     NDArrayObject,
@@ -27,39 +28,6 @@ from floris.utilities import cosd
 
 POWER_SETPOINT_DEFAULT = 1e12
 POWER_SETPOINT_DISABLED = 0.001
-
-
-@define
-class BaseOperationModel(BaseClass):
-    """
-    Base class for turbine operation models. All turbine operation models must implement static
-    power(), thrust_coefficient(), and axial_induction() methods, which are called by power() and
-    thrust_coefficient() through the interface in the turbine.py module.
-
-    Args:
-        BaseClass (_type_): _description_
-
-    Raises:
-        NotImplementedError: _description_
-        NotImplementedError: _description_
-    """
-    @staticmethod
-    @abstractmethod
-    def power() -> None:
-        raise NotImplementedError("BaseOperationModel.power")
-
-    @staticmethod
-    @abstractmethod
-    def thrust_coefficient() -> None:
-        raise NotImplementedError("BaseOperationModel.thrust_coefficient")
-
-    @staticmethod
-    @abstractmethod
-    def axial_induction() -> None:
-        # TODO: Consider whether we can make a generic axial_induction method
-        # based purely on thrust_coefficient so that we don't need to implement
-        # axial_induciton() in individual operation models.
-        raise NotImplementedError("BaseOperationModel.axial_induction")
 
 @define
 class SimpleTurbine(BaseOperationModel):
